# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/integrations": {
      "get": {
        "operationId": "Integrations_list",
        "summary": "Integration log (A12)",
        "description": "Every message in and out, newest first.",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "SUCCESS",
                "FAILED"
              ],
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "IN",
                "OUT"
              ],
              "type": "string"
            }
          },
          {
            "name": "system",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "CRM",
                "ERP",
                "FINANCE",
                "SERVICE",
                "OEM",
                "EMAIL"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated<IntegrationMessage>"
          }
        },
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{id}/retry": {
      "post": {
        "operationId": "Integrations_retry",
        "summary": "Retry a failed message",
        "description": "Sends a FAILED message again (attempts + 1). `409 not_failed` for other messages.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated IntegrationMessage"
          }
        },
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/simulate/job-result": {
      "post": {
        "operationId": "Demo_jobResult",
        "summary": "Demo only: the service system's job result for a WITH_SERVICE complaint (W3)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "complaintId"
                ],
                "properties": {
                  "complaintId": {
                    "type": "string"
                  },
                  "partType": {
                    "type": "string",
                    "enum": [
                      "COMPRESSOR",
                      "PCB"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ComplaintView"
          }
        },
        "tags": [
          "demo"
        ]
      }
    },
    "/api/simulate/oem-decision": {
      "post": {
        "operationId": "Demo_oemDecision",
        "summary": "Demo only: the manufacturer's decision on a SUBMITTED claim (W3)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "claimId",
                  "decision"
                ],
                "properties": {
                  "claimId": {
                    "type": "string"
                  },
                  "decision": {
                    "type": "string",
                    "enum": [
                      "APPROVED",
                      "REJECTED"
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ClaimView"
          }
        },
        "tags": [
          "demo"
        ]
      }
    }
  },
  "info": {
    "title": "Warranty Management API",
    "description": "Units with part-wise warranties, registrations from every channel, complaints, service hand-off, manufacturer claims and the integration log. Contract: frontend/docs/api-contract.md.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wms_refresh"
      }
    },
...
  }
}